% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_values.R
\name{assign_values}
\alias{assign_values}
\title{Assign objects from a function to a list}
\usage{
assign_values(X, function_env, return_list)
}
\arguments{
\item{X}{the object you want to get from the function environment}

\item{function_env}{the environment of the function you are taking the object from}

\item{return_list}{the name of the list you are appending the object to}
}
\value{
the new list with the objects assigned to slots
}
\description{
Function to extract objects from a function environment and append them as slots in a list
}
