% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_assignment.R
\name{generic_assignment}
\alias{generic_assignment}
\title{Assign names and items to list}
\usage{
generic_assignment(list__, element_type, element_value)
}
\arguments{
\item{list__}{the list to append slots to}

\item{element_type}{the name of the list item}

\item{element_value}{the value of the list item}
}
\value{
the list with items
}
\description{
A function to assign new list slots based on an object and value
}
