% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_transformation.R
\name{get_logistic_transform}
\alias{get_logistic_transform}
\title{Get logistic transformation of a parameter}
\usage{
get_logistic_transform(par, lo, hi)
}
\arguments{
\item{par}{the parameter value}

\item{lo}{the low bound}

\item{hi}{the high bound}
}
\value{
the transformed parameter value
}
\description{
Logistically transforms a parameter to bound it between a low and a high value
}
