% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_transformation.R
\name{inv_logistic}
\alias{inv_logistic}
\title{Return logistically transformed variable to normal space}
\usage{
inv_logistic(par, lo, hi)
}
\arguments{
\item{par}{the parameter value}

\item{lo}{the low bound}

\item{hi}{the high bound}
}
\value{
the transformed parameter value
}
\description{
Un-logistically transform a parameter value
}
